+++ rp2040-hal/CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- None
+- Updated embedded-hal alpha support to version 1.0.0-alpha.7
 
 ## [0.3.0] - 2021-12-19